Alphabetize header links
G comes before H. Novel, I know.

This is an example of using judgement about when a fix belongs in a pull
request or needs to be split out into its own pull request. This
particular case is a trivial change. It's not user-visible (these
footnotes aren't shown in the rendered document), so there's no hurry to
get it shipped. This pull request already touches the footnotes, so
having this change in the same pull request won't distract from its
focus.

I did, however, put it in its own commit so that when I add more links
here, those commits will only contain added lines. It'll be much easier
to review two commits that sort lines and then add lines. Doing it as a
single commit would force the reviewer to trace removed lines and make
sure they're added back in order elsewhere, and if there were a bunch of
these changes, we might accidentally delete one and sneak it through.

I disocvered this as I was working on the third commit in this branch,
"Tell a linear story". To avoid distracting the reviewer with a
non-etiquette change in the middle of the pull request, I used `git
rebase --interactive` to move this commit to the beginning of the
branch. I do this often with refactor commits that fit in the branch but
aren't part of the core logic of the feature I'm trying to implement.